requirements.txt
# ====================================
# Backend Dependencies
# ====================================

# FastAPI - Web Framework
fastapi==0.127.0
uvicorn[standard]==0.40.0

# Supabase Client
supabase==2.27.0
postgrest==2.27.0
gotrue==2.12.4

# HTTP Client
httpx==0.28.1

# Data Validation
pydantic==2.12.5
pydantic-settings==2.12.0

# Environment Variables
python-dotenv==1.2.1

# ====================================
# Data Processing (for scripts)
# ====================================

# Excel/CSV handling
pandas==2.3.3
openpyxl==3.1.5

# ====================================
# Notes
# ====================================
# Compatible with Python 3.11 - 3.12
# Install with: pip install -r requirements.txt


